refactor: report state to its own helper script
